Fixed project logic

+++ frictionless/project/project.py
@@ -69,29 +69,56 @@ class Project:
         name = os.path.basename(path)
         folder = folder or os.path.dirname(path)
         source = str(self.public / path)
+        basetarget = str(self.public / folder)
         target = str(self.public / folder / name)
         target = deduplicate_path(target, suffix="copy")
-        assert os.path.isfile(source)
-        helpers.copy_file(source, target)
+        assert os.path.isdir(basetarget)
+        # File
+        if os.path.isfile(source):
+            helpers.copy_file(source, target)
+        # Folder
+        elif os.path.isdir(source):
+            helpers.copy_folder(source, target)
+        # Missing
+        else:
+            raise FrictionlessException("file doesn't exist")
         path = str(Path(target).relative_to(self.public))
         return path
 
-    # TODO: use folder?
     # TODO: use streaming?
     def file_create(
         self, name: str, *, bytes: bytes, folder: Optional[str] = None
     ) -> str:
         assert not os.path.dirname(name)
+        basepath = str(self.public / (folder or ""))
         path = str(self.public / (folder or "") / name)
         path = deduplicate_path(path)
+        assert os.path.isdir(basepath)
         helpers.write_file(path, bytes, mode="wb")
+        path = str(Path(path).relative_to(self.public))
+        return path
+
+    def file_create_folder(self, name: str, *, folder: Optional[str] = None) -> str:
+        assert not os.path.dirname(name)
+        basepath = str(self.public / (folder or ""))
+        path = str(self.public / (folder or "") / name)
+        assert os.path.isdir(basepath)
+        assert not os.path.exists(path)
+        Path(path).mkdir(parents=True, exist_ok=False)
         path = str(Path(path).relative_to(self.public))
         return path
 
     def file_delete(self, path: str) -> str:
         path = str(self.public / path)
-        assert os.path.isfile(path)
-        os.remove(path)
+        # File
+        if os.path.isfile(path):
+            os.remove(path)
+        # Folder
+        elif os.path.isdir(path):
+            shutil.rmtree(path)
+        # Missing
+        else:
+            FrictionlessException("file doesn't exist")
         path = str(Path(path).relative_to(self.public))
         return path
 
@@ -128,10 +155,20 @@ class Project:
     def file_move(self, path: str, *, folder: str) -> str:
         name = os.path.basename(path)
         source = str(self.public / path)
+        basetarget = str(self.public / folder)
         target = str(self.public / folder / name)
-        assert os.path.isfile(source)
+        target = deduplicate_path(target)
+        assert os.path.isdir(basetarget)
         assert not os.path.exists(target)
-        helpers.move_file(source, target)
+        # File
+        if os.path.isfile(source):
+            helpers.move_file(source, target)
+        # Folder
+        elif os.path.isdir(source):
+            helpers.move_folder(source, target)
+        # Missing
+        else:
+            raise FrictionlessException("file doesn't exist")
         path = str(Path(target).relative_to(self.public))
         return path
 
@@ -141,41 +178,6 @@ class Project:
         assert os.path.isfile(path)
         bytes = helpers.read_file(path, "rb")
         return bytes
-
-    # Folder
-
-    def folder_copy(self, path: str, *, folder: Optional[str] = None) -> str:
-        folder = folder or os.path.dirname(path)
-        source = str(self.public / path)
-        target = str(self.public / folder)
-        target = deduplicate_path(target)
-        assert os.path.isdir(source)
-        helpers.copy_folder(source, target)
-        path = str(Path(target).relative_to(self.public))
-        return path
-
-    def folder_create(self, path: str) -> str:
-        path = str(self.public / path)
-        assert not os.path.exists(path)
-        Path(path).mkdir(parents=True, exist_ok=True)
-        path = str(Path(path).relative_to(self.public))
-        return path
-
-    def folder_delete(self, path: str) -> str:
-        path = str(self.public / path)
-        assert os.path.isdir(path)
-        shutil.rmtree(path)
-        path = str(Path(path).relative_to(self.public))
-        return path
-
-    def folder_move(self, path: str, *, folder: str) -> str:
-        source = str(self.public / path)
-        target = str(self.public / folder)
-        assert os.path.isdir(source)
-        assert not os.path.exists(target)
-        shutil.move(source, target)
-        path = str(Path(target).relative_to(self.public))
-        return path
 
     # Package
